chore: migrate events to TS
